+++ src/Uccelli.java
@@ -1,7 +1,7 @@
 
 public abstract class Uccelli extends Animale {
 
-	public Uccelli(int vel, int f, int vit, int ene, int att, int dif, int da, int alt, int lung, int larg, double pe) {
+	public Uccelli(double vel, int f, int vit, int ene, int att, int dif, int da, int alt, int lung, int larg, double pe) {
 		super(vel, f, vit, ene, att, dif, da, alt, lung, larg, pe);
 		
 	}
